requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0
django-phonenumber-field==7.3.0
django-twilio==0.14.3.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environs==11.0.0
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
marshmallow==3.21.1
multidict==6.0.5
packaging==24.0
phonenumbers==8.13.34
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
twilio==9.0.4
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
asgiref==3.8.1
Django
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
